// src/Screens/MerchantDetails/types.ts
export type AboutSectionHeadingContentContainerProps = {
    alignItems: string;
    justifyContent?: string;
    width?: string;
};

export type CarouselContainerProps = {
    height?: number;
    width?: number;
};

export type CarouselImageProps = {
    height?: number;
    width?: number;
};

export type CarouselPaginationDotProps = {
    active?: boolean;
};
   
export type ProductCardContentProps = {
    width: string;
    alignItems?: string;
    paddingRight?: number;
};

export type ProductCardContainerType = {
    backgroundColor?: string;
};